+++ src/python/midi.py
@@ -1,47 +1,5 @@
-from sys import argv, stdin, stdout, version_info, version
-from binascii import hexlify, unhexlify
-import traceback
-import json as JSON
-import re
-
-def ipcSend(name, data):
-    print(JSON.dumps([name, data]))
-    stdout.flush()
-
-try:
-    import rtmidi
-    API = rtmidi.RtMidiIn.UNIX_JACK if 'jack' in argv else rtmidi.RtMidiIn.UNSPECIFIED
-    JACK = API == rtmidi.RtMidiIn.UNIX_JACK
-    in_dev = rtmidi.RtMidiIn(API, 'MIDI->OSC probe')
-    out_dev = rtmidi.RtMidiOut(API, 'OSC->MIDI probe')
-except:
-    ipcSend('error', 'pyrtmidi not found (or wrong version)\nRunning with python version %s' % version)
-
-
-if version_info.major == 3:
-    raw_input = input
-
-debug = False
-if 'debug' in argv:
-    debug = True
-
-def list():
-
-    message = []
-
-    message.append('===========')
-    message.append('MIDI Inputs')
-
-    for i in range(in_dev.getPortCount()):
-        message.append('%i: %s' % (i, in_dev.getPortName(i)))
-
-    message.append('===========')
-    message.append('MIDI Outputs')
-
-    for i in range(out_dev.getPortCount()):
-        message.append('%i: %s' % (i, out_dev.getPortName(i)))
-
-    ipcSend('log', '\n'.join(message))
+from head import *
+from list import *
 
 if 'list' in argv:
     list()