+++ src/server/index.js
@@ -98,6 +98,7 @@ if (settings.cli) {
 
     app.on('ready',function(){
         global.settings = settings
+        global.midilist = require('./midi').list
         launcher = require('./electron-window')({address:address, shortcuts:dev, width:680, height:(100 + 8*3 + 29 * Object.keys(settings.options).filter(x=>settings.options[x].launcher !== false).length), node:true, color:'#253040'})
     })