+++ src/client/app/widgets/inputs/keys.js
@@ -16,6 +16,7 @@ class Keys extends Widget {
             keydown: {type: 'string', value: '', help: [
                 'This property is evaluated each time the key combo is pressed and defines the widget\'s own value. Formulas are given extras variables in this context:',
                 '- `key`: pressed key name (usefull for handling multiple keys with a single keys widget, lowercased when referencing a character key)',
+                '- `code`: pressed key code name (<a href="https://developer.mozilla.org/en-US/docs/Web/API/KeyboardEvent/code#Code_values" target="_blank">full list</a>)',
                 '- `ctrl`: control key state',
                 '- `alt`: alt key state',
                 '- `shift`: shift key state',
@@ -80,6 +81,7 @@ class Keys extends Widget {
 
             var context = {
                 key: e.key,
+                code: e.code,
                 ctrl: e.ctrlKey,
                 shift: e.shiftKey,
                 alt: e.altKey,
@@ -108,6 +110,7 @@ class Keys extends Widget {
 
             var context = {
                 key: e.key,
+                code: e.code,
                 ctrl: e.ctrlKey,
                 shift: e.shiftKey,
                 alt: e.altKey,